+++ src/stac_auth_proxy/middleware/UpdateOpenApiMiddleware.py
